+++ confiture-web-app/src/pages/edit/NewAuditStepOnePage.vue
@@ -76,7 +76,9 @@ async function submitStepOne(data: CreateAuditRequestData) {
   auditStore
     .createAudit(data)
     .then((audit) => {
-      auditStore.showAuditEmailAlert = true;
+      if (!accountStore.account) {
+        auditStore.showAuditEmailAlert = true;
+      }
       // TODO: replace current history entry with the edit page
       return router.push({
         name: "overview",
